# openapi.yaml
openapi: 3.0.3
info:
  title: Feedback Platform API
  version: 1.0.0
  description: |
    API serverless do projeto feedback-platform-core exposta pelo Amazon API Gateway e protegida por Amazon Cognito.

    Boas praticas aplicadas neste contrato:
    - autenticacao centralizada no API Gateway com bearer JWT
    - schemas de request e response versionados no contrato
    - exemplos de payload para teste rapido
    - respostas de erro padronizadas para 4xx e 5xx
servers:
  - url: https://r2iu71nvl4.execute-api.us-east-2.amazonaws.com/dev
    description: Ambiente dev publicado na AWS
tags:
  - name: Feedback
    description: Coleta e consulta de feedbacks
  - name: Reports
    description: Consultas analiticas e relatorios
security:
  - CognitoOAuth2:
      - openid
paths:
  /avaliacao:
    post:
      tags:
        - Feedback
      summary: Registrar avaliacao
      description: Recebe uma avaliacao no formato simplificado do enunciado.
      operationId: createAvaliacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvaliacaoRequest'
            examples:
              default:
                value:
                  descricao: Aula clara e objetiva.
                  nota: 9
      responses:
        '201':
          description: Avaliacao recebida com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Feedback recebido com sucesso
        '400':
          $ref: '#/components/responses/BadRequestMessage'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerErrorMessage'
      x-amazon-apigateway-request-validator: body-only
  /feedback:
    post:
      tags:
        - Feedback
      summary: Registrar feedback
      description: Recebe um feedback academico e encaminha o processamento pela Lambda de coleta.
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
            examples:
              default:
                value:
                  cursoId: 1TIA
                  alunoId: aluno-001
                  professorId: prof-123
                  nota: 9
                  comentario: Aula clara, com boa explicacao dos exercicios.
      responses:
        '201':
          description: Feedback recebido com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Feedback recebido com sucesso
        '400':
          $ref: '#/components/responses/BadRequestMessage'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerErrorMessage'
      x-amazon-apigateway-request-validator: body-only
  /feedback/{id}:
    get:
      tags:
        - Feedback
      summary: Consultar feedback por ID
      description: Recupera um feedback previamente armazenado pelo identificador informado.
      operationId: getFeedbackById
      parameters:
        - $ref: '#/components/parameters/FeedbackId'
      responses:
        '200':
          description: Feedback encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
              example:
                id: fb-1a2b3c
                cursoId: 1TIA
                alunoId: aluno-001
                professorId: prof-123
                nota: 9
                comentario: Aula clara, com boa explicacao dos exercicios.
                criticidade: BAIXA
                dataCriacao: '2026-07-12T13:15:10Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/professor/{professorId}:
    get:
      tags:
        - Reports
      summary: Consultar relatorio por professor
      description: Retorna os feedbacks agregados para um professor.
      operationId: getProfessorReport
      parameters:
        - $ref: '#/components/parameters/ProfessorId'
      responses:
        '200':
          description: Relatorio retornado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfessorReportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/curso/{cursoId}:
    get:
      tags:
        - Reports
      summary: Consultar relatorio por curso
      description: Retorna os feedbacks agregados para um curso.
      operationId: getCourseReport
      parameters:
        - $ref: '#/components/parameters/CursoId'
      responses:
        '200':
          description: Relatorio retornado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursoReportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/semanal:
    get:
      tags:
        - Reports
      summary: Consultar relatorio semanal por curso
      description: Gera um consolidado semanal por curso, opcionalmente filtrando por professor.
      operationId: getWeeklyCourseReport
      parameters:
        - name: cursoId
          in: query
          required: true
          description: Identificador do curso.
          schema:
            type: string
        - name: professorId
          in: query
          required: false
          description: Identificador do professor para filtragem opcional.
          schema:
            type: string
      responses:
        '200':
          description: Relatorio semanal retornado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyCourseReportResponse'
        '400':
          description: Parametros invalidos.
          content:
            text/plain:
              schema:
                type: string
              example: cursoId é obrigatório
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    CognitoOAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://feedback-platform-022267198187-us-east-2.auth.us-east-2.amazoncognito.com/oauth2/authorize
          tokenUrl: https://feedback-platform-022267198187-us-east-2.auth.us-east-2.amazoncognito.com/oauth2/token
          scopes:
            openid: Escopo padrao para autenticacao via Cognito.
      description: |
        Fluxo OAuth2 Authorization Code via Amazon Cognito para uso do botao Authorize no Swagger UI.
  parameters:
    FeedbackId:
      name: id
      in: path
      required: true
      description: Identificador unico do feedback.
      schema:
        type: string
    ProfessorId:
      name: professorId
      in: path
      required: true
      description: Identificador do professor.
      schema:
        type: string
    CursoId:
      name: cursoId
      in: path
      required: true
      description: Identificador do curso.
      schema:
        type: string
  responses:
    BadRequestMessage:
      description: Requisicao invalida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Payload de entrada é obrigatório
    UnauthorizedError:
      description: Token ausente, invalido ou expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
          example:
            error_code: TOKEN_INVALID
            message: Authorization Bearer token is required
            trace_id: 2d7ddf8d-9c23-4b3a-aa1f-3fe0e4d9b111
    ForbiddenError:
      description: Token valido sem permissao suficiente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
          example:
            error_code: INSUFFICIENT_PERMISSION
            message: JWT must contain at least one role in 'groups' or 'scope'
            trace_id: 2d7ddf8d-9c23-4b3a-aa1f-3fe0e4d9b111
    NotFoundError:
      description: Recurso nao encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
          example:
            error_code: NOT_FOUND
            message: Feedback nao encontrado
            trace_id: 2d7ddf8d-9c23-4b3a-aa1f-3fe0e4d9b111
    InternalServerError:
      description: Erro interno no processamento.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessageResponse'
          example:
            error: Erro ao processar notificação
    InternalServerErrorMessage:
      description: Erro interno no processamento.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Erro interno ao processar feedback
  schemas:
    AvaliacaoRequest:
      type: object
      required:
        - descricao
        - nota
      properties:
        descricao:
          type: string
        nota:
          type: integer
          minimum: 0
          maximum: 10
      additionalProperties: false
    FeedbackRequest:
      type: object
      required:
        - cursoId
        - alunoId
        - professorId
        - nota
        - comentario
      properties:
        cursoId:
          type: string
        alunoId:
          type: string
        professorId:
          type: string
        nota:
          type: integer
          minimum: 0
          maximum: 10
        comentario:
          type: string
      additionalProperties: false
    FeedbackResponse:
      type: object
      properties:
        id:
          type: string
        cursoId:
          type: string
        alunoId:
          type: string
        professorId:
          type: string
        nota:
          type: integer
        comentario:
          type: string
        criticidade:
          type: string
          enum:
            - BAIXA
            - MEDIA
            - ALTA
            - CRITICA
        dataCriacao:
          type: string
          format: date-time
    FeedbackReportItem:
      type: object
      properties:
        id:
          type: string
        cursoId:
          type: string
        alunoId:
          type: string
        professorId:
          type: string
        nota:
          type: integer
          nullable: true
        comentario:
          type: string
        criticidade:
          type: string
        dataCriacao:
          type: string
          format: date-time
    ProfessorReportResponse:
      type: object
      properties:
        professorId:
          type: string
        totalFeedbacks:
          type: integer
        feedbacks:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackReportItem'
    CursoReportResponse:
      type: object
      properties:
        cursoId:
          type: string
        totalFeedbacks:
          type: integer
        feedbacks:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackReportItem'
    WeeklyCourseReportResponse:
      type: object
      properties:
        cursoId:
          type: string
        professorId:
          type: string
          nullable: true
        totalFeedbacks:
          type: integer
        mediaNota:
          type: number
          format: double
        quantidadeBaixa:
          type: integer
          format: int64
        quantidadeMedia:
          type: integer
          format: int64
        quantidadeAlta:
          type: integer
          format: int64
        feedbacksPorProfessor:
          type: object
          additionalProperties:
            type: integer
            format: int64
        geradoEm:
          type: string
          format: date-time
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ErrorMessageResponse:
      type: object
      properties:
        error:
          type: string
    AuthErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
        trace_id:
          type: string
x-amazon-apigateway-request-validators:
  body-only:
    validateRequestBody: true
    validateRequestParameters: false